requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# E:\NAM-IP\catTools-git\catTools_app\init_db.py: 3
# E:\NAM-IP\catTools-git\catTools_app\migration_BULL\outils_bull\db.py: 2
# E:\NAM-IP\catTools-git\catTools_app\migration_BULL\proto_bull.py: 7
# E:\NAM-IP\catTools-git\catTools_app\migration_KHS\outils_khs.py: 8,9
# E:\NAM-IP\catTools-git\catTools_app\migration_KHS\proto_khs.py: 5
# E:\NAM-IP\catTools-git\dB_lieux\remplir_db.py: 1
# E:\NAM-IP\catTools-git\run.py: 1
mysql_connector_python == 8.0.28

# E:\NAM-IP\catTools-git\dB_lieux\remplir_db.py: 2
numpy == 1.22.2

# E:\NAM-IP\catTools-git\catTools_app\migration_BULL\proto_bull.py: 6
openpyxl == 3.0.9

# E:\NAM-IP\catTools-git\dB_lieux\remplir_db.py: 4
# E:\NAM-IP\catTools-git\test.py: 1
pandas == 1.4.1

# E:\NAM-IP\catTools-git\dB_lieux\remplir_db.py: 7
tqdm == 4.64.0

# E:\NAM-IP\catTools-git\dB_lieux\remplir_db.py: 3
translate == 3.6.1
